telemetry typings as top component and helpers moved into platform
