+++ src/notebooks/telemetry/interpreterPackageTracker.ts
@@ -12,7 +12,7 @@ import { INotebookControllerManager } from '../types';
 import { IInstaller, Product } from '../../kernels/installer/types';
 import { IVSCodeNotebookController } from '../controllers/types';
 import { trackKernelResourceInformation } from '../../kernels/telemetry/helper';
-import { IInterpreterPackages } from '../../telemetryTypes';
+import { IInterpreterPackages } from '../../telemetry';
 
 @injectable()
 export class InterpreterPackageTracker implements IExtensionSingleActivationService {
