+++ src/notebooks/telemetry/notebookOrKernelLanguageTelemetry.ts
@@ -3,7 +3,7 @@
 'use strict';
 
 import { sendTelemetryEvent, Telemetry } from '../../telemetry';
-import { getTelemetrySafeLanguage } from '../../telemetry/helpers';
+import { getTelemetrySafeLanguage } from '../../platform/telemetry/helpers';
 
 export function sendNotebookOrKernelLanguageTelemetry(
     telemetryEvent: Telemetry.SwitchToExistingKernel | Telemetry.NotebookLanguage,
