+++ src/platform/common/variables/environmentVariablesProvider.node.ts
@@ -4,7 +4,7 @@ import { inject, injectable, optional } from 'inversify';
 import * as path from '../../vscode-path/path';
 import { ConfigurationChangeEvent, Disposable, Event, EventEmitter, FileSystemWatcher, Uri } from 'vscode';
 import { TraceOptions } from '../../logging/types';
-import { sendFileCreationTelemetry } from '../../../telemetry/envFileTelemetry.node';
+import { sendFileCreationTelemetry } from '../../telemetry/envFileTelemetry.node';
 import { IWorkspaceService } from '../application/types';
 import { IDisposableRegistry } from '../types';
 import { InMemoryCache } from '../utils/cacheUtils';
