+++ src/platform/telemetry/constants.ts
