+++ src/platform/telemetry/extensionInstallTelemetry.node.ts
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import * as path from '../platform/vscode-path/path';
+import * as path from '../vscode-path/path';
 import { setSharedProperty } from '.';
-import { IFileSystemNode } from '../platform/common/platform/types.node';
-import { EXTENSION_ROOT_DIR } from '../platform/constants.node';
+import { IFileSystemNode } from '../common/platform/types.node';
+import { EXTENSION_ROOT_DIR } from '../constants.node';
 
 /**
  * Sets shared telemetry property about where the extension was installed from
