+++ src/platform/telemetry/helpers.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 
 import * as hashjs from 'hash.js';
-import { KnownKernelLanguageAliases, VSCodeKnownNotebookLanguages } from '../platform/common/constants';
-import { traceError } from '../platform/logging';
+import { KnownKernelLanguageAliases, VSCodeKnownNotebookLanguages } from '../common/constants';
+import { traceError } from '../logging';
 
 export function getTelemetrySafeLanguage(language: string = 'unknown') {
     language = (language || 'unknown').toLowerCase();
