+++ src/platform/telemetry/interpreterPackages.node.ts
@@ -2,16 +2,16 @@
 // Licensed under the MIT License.
 
 import { inject, injectable } from 'inversify';
-import { IPythonApiProvider, IPythonExtensionChecker } from '../platform/api/types';
-import { IPythonExecutionFactory } from '../platform/common/process/types.node';
-import { IDisposableRegistry, InterpreterUri } from '../platform/common/types';
-import { createDeferred, Deferred } from '../platform/common/utils/async';
-import { isResource, noop } from '../platform/common/utils/misc';
-import { IInterpreterService } from '../platform/interpreter/contracts';
-import { PythonEnvironment } from '../platform/pythonEnvironments/info';
-import { getComparisonKey } from '../platform/vscode-path/resources';
+import { IPythonApiProvider, IPythonExtensionChecker } from '../api/types';
+import { IPythonExecutionFactory } from '../common/process/types.node';
+import { IDisposableRegistry, InterpreterUri } from '../common/types';
+import { createDeferred, Deferred } from '../common/utils/async';
+import { isResource, noop } from '../common/utils/misc';
+import { IInterpreterService } from '../interpreter/contracts';
+import { PythonEnvironment } from '../pythonEnvironments/info';
+import { getComparisonKey } from '../vscode-path/resources';
 import { getTelemetrySafeHashedString, getTelemetrySafeVersion } from './helpers';
-import { IInterpreterPackages } from '../telemetryTypes';
+import { IInterpreterPackages } from '../../telemetry';
 
 const interestedPackages = new Set(
     [
