+++ src/standalone/activation/workspaceActivation.node.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from 'inversify';
 import { TextDocument } from 'vscode';
-import { sendActivationTelemetry } from '../../telemetry/envFileTelemetry.node';
+import { sendActivationTelemetry } from '../../platform/telemetry/envFileTelemetry.node';
 import { IPythonExtensionChecker } from '../../platform/api/types';
 import { IWorkspaceService, IActiveResourceService, IDocumentManager } from '../../platform/common/application/types';
 import { PYTHON_LANGUAGE } from '../../platform/common/constants';
