+++ src/test/common/experiments/service.unit.test.ts
@@ -14,7 +14,7 @@ import { ConfigurationService } from '../../../platform/common/configuration/ser
 import { ExperimentService } from '../../../platform/common/experiments/service';
 import { IConfigurationService } from '../../../platform/common/types';
 import * as Telemetry from '../../../telemetry';
-import { EventName } from '../../../telemetry/constants';
+import { EventName } from '../../../platform/telemetry/constants';
 import { JVSC_EXTENSION_ID_FOR_TESTS } from '../../constants.node';
 import { MockOutputChannel } from '../../mockClasses';
 import { MockMemento } from '../../mocks/mementos';
