+++ src/test/kernels/installer/productPath.unit.test.ts
@@ -10,7 +10,7 @@ import * as TypeMoq from 'typemoq';
 import { Memento, Uri } from 'vscode';
 import '../../../platform/common/extensions';
 import { IConfigurationService, IOutputChannel } from '../../../platform/common/types';
-import { InterpreterPackages } from '../../../telemetry/interpreterPackages.node';
+import { InterpreterPackages } from '../../../platform/telemetry/interpreterPackages.node';
 import { IServiceContainer } from '../../../platform/ioc/types';
 import { ProductInstaller } from '../../../kernels/installer/productInstaller.node';
 import { BaseProductPathsService } from '../../../kernels/installer/productPath.node';
