+++ src/test/telemetry/importTracker.unit.test.ts
@@ -17,8 +17,8 @@ import {
 } from '../../platform/common/constants';
 import { disposeAllDisposables } from '../../platform/common/helpers';
 import { IDisposable } from '../../platform/common/types';
-import { EventName } from '../../telemetry/constants';
-import { getTelemetrySafeHashedString } from '../../telemetry/helpers';
+import { EventName } from '../../platform/telemetry/constants';
+import { getTelemetrySafeHashedString } from '../../platform/telemetry/helpers';
 import { ImportTracker } from '../../standalone/import-export/importTracker.node';
 import { createDocument } from '../datascience/editor-integration/helpers';
